Moved most code from HDRP to VFX. VFX dependent and has access to Shadergraph internals.
