+++ com.unity.render-pipelines.high-definition/Runtime/Sky/PhysicallyBasedSky.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: cdf7c1bc75c42c648afb3f04b2f88617
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
