+++ com.unity.shadergraph/Editor/AssemblyInfo.cs
@@ -5,3 +5,4 @@ using System.Runtime.CompilerServices;
 [assembly: InternalsVisibleTo("Unity.ShaderGraph.Editor.GraphicsTests")]
 [assembly: InternalsVisibleTo("Unity.RenderPipelines.Lightweight.Editor")]
 [assembly: InternalsVisibleTo("Unity.RenderPipelines.HighDefinition.Editor")]
+[assembly: InternalsVisibleTo("Unity.VisualEffectGraph.Editor")]
