+++ com.unity.visualeffectgraph/Editor/ShaderGraph.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: bb0537d0b085782498bcb7d34b0bd86c
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
