+++ com.unity.visualeffectgraph/Editor/ShaderGraph/Output.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: c033f6abbe05ccb45b1baa59bc20076c
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
