+++ com.unity.visualeffectgraph/Editor/ShaderGraph/Output/VFXMeshShaderGraphOutput.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: f4476dfd621f4fe428fe6e72d10aa015
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
