+++ com.unity.visualeffectgraph/Editor/ShaderGraph/Output/VFXShaderGraphOutput.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: d794971dca291b241b31d94cede164d2
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
