+++ com.unity.visualeffectgraph/Editor/ShaderGraph/ShaderParser/SubShaderPart.cs
@@ -2,7 +2,7 @@ using System.Collections.Generic;
 using UnityEngine;
 using UnityEditor.ShaderGraph;
 
-namespace UnityEditor.Experimental.Rendering.HDPipeline.VFXSG
+namespace UnityEditor.VFX.SG
 {
     class SubShaderPart : ShaderPart
     {
